Improve fundamental analysis files organization
