+++ gamestonk_terminal/fundamental_analysis/fa_controller.py
@@ -5,11 +5,11 @@ import argparse
 from typing import List
 from prompt_toolkit.completion import NestedCompleter
 
-from gamestonk_terminal.fundamental_analysis import alpha_vantage_controller as avc
-from gamestonk_terminal.fundamental_analysis import business_insider_view as biw
-from gamestonk_terminal.fundamental_analysis import (
-    financial_modeling_prep_controller as fmpc,
-    financial_modeling_prep_view as fmpv,
+from gamestonk_terminal.fundamental_analysis.alpha_vantage import av_controller
+from gamestonk_terminal.fundamental_analysis import business_insider_view
+from gamestonk_terminal.fundamental_analysis.financial_modeling_prep import (
+    fmp_controller,
+    fmp_view,
 )
 from gamestonk_terminal.fundamental_analysis import finviz_view
 from gamestonk_terminal.fundamental_analysis import market_watch_view
@@ -136,7 +136,7 @@ class FundamentalAnalysisController:
 
     def call_mgmt(self, other_args: List[str]):
         """Process mgmt command"""
-        biw.management(other_args, self.ticker)
+        business_insider_view.management(other_args, self.ticker)
 
     def call_screener(self, other_args: List[str]):
         """Process screener command"""
@@ -144,7 +144,7 @@ class FundamentalAnalysisController:
 
     def call_score(self, other_args: List[str]):
         """Process score command"""
-        fmpv.valinvest_score(other_args, self.ticker)
+        fmp_view.valinvest_score(other_args, self.ticker)
 
     def call_income(self, other_args: List[str]):
         """Process income command"""
@@ -174,18 +174,16 @@ class FundamentalAnalysisController:
         """Process cal command"""
         yahoo_finance_view.calendar_earnings(other_args, self.ticker)
 
-    # pylint: disable=unused-argument
-    def call_av(self, other_args: List[str]):
+    def call_av(self, _):
         """Process av command"""
-        ret = avc.menu(self.ticker, self.start, self.interval)
+        ret = av_controller.menu(self.ticker, self.start, self.interval)
 
         if ret is not True:
             self.print_help()
 
-    # pylint: disable=unused-argument
-    def call_fmp(self, other_args: List[str]):
+    def call_fmp(self, _):
         """Process fmp command"""
-        ret = fmpc.menu(self.ticker, self.start, self.interval)
+        ret = fmp_controller.menu(self.ticker, self.start, self.interval)
 
         if ret is not True:
             self.print_help()
