+++ gamestonk_terminal/fundamental_analysis/financial_modeling_prep/fmp_view.py
